split _init into __init__ and _from_numpy, and add stub
